Factor out shader copy into own class
 Factors out the shader copying code into its own class, ShaderCopy.
 This will allow us to create separate implementations to perform the
 copy that do not just use shaders.
